third_party/WebKit/LayoutTests/http/tests/storage/async_cookies/cookieStore_delete_arguments.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookes: cookieStore.delete() arguments and options</title>
<link rel="author" href="pwnall@chromium.org" title="Victor Costan">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');

  await cookieStore.delete('cookie-name');
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie, null);
}, 'cookieStore.delete with positional name');

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');

  await cookieStore.delete({ name: 'cookie-name' });
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie, null);

  try { await cookieStore.delete('cookie-name'); } catch (e) {}
}, 'cookieStore.delete with name in options');

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');

  await promise_rejects(testCase, new TypeError(), cookieStore.delete(
      'cookie-name', { name: 'cookie-name' }));
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie.name, 'cookie-name');
  assert_equals(cookie.value, 'cookie-value');

  try { await cookieStore.delete('cookie-name'); } catch (e) {}
}, 'cookieStore.set with name in both positional arguments and options');

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');

  await promise_rejects(testCase, new TypeError(), cookieStore.delete(
      'cookie-name', { value: 'cookie-value' }));
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie.name, 'cookie-name');
  assert_equals(cookie.value, 'cookie-value');

  try { await cookieStore.delete('cookie-name'); } catch (e) {}
}, 'cookieStore.delete with value in options');

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');

  const tenYears = 10 * 365 * 24 * 60 * 60 * 1000;
  const tenYearsAgo = Date.now() - tenYears;

  await promise_rejects(testCase, new TypeError(), cookieStore.delete(
      'cookie-name', { expires: tenYearsAgo }));
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie.name, 'cookie-name');
  assert_equals(cookie.value, 'cookie-value');

  try { await cookieStore.delete('cookie-name'); } catch (e) {}
}, 'cookieStore.delete with expires in options');

promise_test(async testCase => {
  const currentUrl = new URL(window.location.href);
  const currentDomain = currentUrl.hostname;
  await cookieStore.set(
      'cookie-name', 'cookie-value', { domain: currentDomain });

  await cookieStore.delete('cookie-name', { domain: currentDomain });
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie, null);

  try {
    await cookieStore.delete('cookie-name', { domain: currentDomain });
  } catch (e) {}
}, 'cookieStore.delete with domain set to the current hostname');

promise_test(async testCase => {
  const currentUrl = new URL(window.location.href);
  const currentDomain = currentUrl.hostname;
  const subDomain = `sub.${currentDomain}`;
  await cookieStore.set(
      'cookie-name', 'cookie-value', { domain: currentDomain });

  await cookieStore.delete('cookie-name', { domain: subDomain });
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie.name, 'cookie-name');
  assert_equals(cookie.value, 'cookie-value');

  try {
    await cookieStore.delete('cookie-name', { domain: currentDomain });
  } catch (e) {}
}, 'cookieStore.delete with domain set to a subdomain of the current hostname');

promise_test(async testCase => {
  const currentUrl = new URL(window.location.href);
  const currentPath = currentUrl.pathname;
  const currentDirectory =
      currentPath.substr(0, currentPath.lastIndexOf('/') + 1);
  await cookieStore.set(
      'cookie-name', 'cookie-value', { path: currentDirectory });

  await cookieStore.delete('cookie-name', { path: currentDirectory });
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie, null);

  try {
    await cookieStore.delete('cookie-name', { path: currentDirectory });
  } catch (e) {}
}, 'cookieStore.delete with path set to the current directory');

promise_test(async testCase => {
  const currentUrl = new URL(window.location.href);
  const currentPath = currentUrl.pathname;
  const currentDirectory =
      currentPath.substr(0, currentPath.lastIndexOf('/') + 1);
  const subDirectory = currentDirectory + "subdir/";
  await cookieStore.set(
      'cookie-name', 'cookie-value', { path: currentDirectory });

  await cookieStore.delete('cookie-name', { path: subDirectory });
  const cookie = await cookieStore.get('cookie-name');
  assert_equals(cookie.name, 'cookie-name');
  assert_equals(cookie.value, 'cookie-value');

  try {
    await cookieStore.delete('cookie-name', { path: currentDirectory });
  } catch (e) {}
}, 'cookieStore.delete with path set to subdirectory of the current directory');

</script>
